requirements.txt
torch
numpy
Pillow
websockets 